# backend/src/main/java/com/matrix/ai/dto/request/RegisterRequest.java
package com.matrix.ai.dto.request;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

/**
 * 用户注册请求
 */
@Data
public class RegisterRequest {

    /**
     * 用户名
     */
    @NotBlank(message = "用户名不能为空")
    @Size(min = 3, max = 50, message = "用户名长度必须在 3-50 个字符之间")
    private String username;

    /**
     * 密码
     */
    @NotBlank(message = "密码不能为空")
    @Size(min = 6, max = 50, message = "密码长度必须在 6-50 个字符之间")
    private String password;

    /**
     * 昵称（可选）
     */
    @Size(max = 50, message = "昵称长度不能超过 50 个字符")
    private String nickname;
}