# backend/src/main/java/com/matrix/ai/dto/response/AuthResponse.java
package com.matrix.ai.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * 认证响应（登录/注册返回）
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {

    /**
     * 用户 ID
     */
    private String userId;

    /**
     * 用户名
     */
    private String username;

    /**
     * 昵称
     */
    private String nickname;

    /**
     * Access Token
     */
    private String token;

    /**
     * Token 有效期（秒）
     */
    private Long expiresIn;
}